Moved to NSFetchedResultsController, image resizing, central error logging
To get rid of the memory issues, we save an image only as big as the screen. An uncompressed image of the size the iPhone4 takes is on the order of 8MB.
GalleryViewController now has two MealViewControllers, and it switches back and forth between them. Previously, you'd see the new page appear, and right at that same time the old page would go blank.
We now use NSFetchedResultsController instead of querying Core Data and saving the results ourselves. It watches the database for us, letting us know when data changes. It means that we now support the scenario where you take a picture then immediately scroll back one page. NSFetchedResultsController is mean to back a UITableView with sections and rows, so we work around that by working entirely within Section 0.
AppDelegate::logError added. In the future, this will be useful as the central place we log errors - whether we want to kill the app or to send it to our analytics provider.
Deleted the Master and Detail views, as well as any references to them.
Further cleanup - killing unnecessary comments and methods. Moving some methods around to keep semantically-similar functions together.