A number of fixes, added RateMealController
Hid the status bar
Fixed first-display issue with MealViewController - setModel was being called before viewDidLoad, which meant that appDelegate wasn't initialized. Interesting to note that when working with PageViewController, any calls you make in after/beforeViewController happen before viewDidLoad or viewWillAppear. :0(
Removed the black bar across the top of the application - this was because the PageViewController wasn't big enough inside its parent view, GalleryViewController.
Skeleton of RateMealController was added.