Show rating on MealView, fix to rendering issue on startup, saving ratings
When the app starts up, it moves to the final page in the scrollview, showing the camera. As it does that, it loads up the meals right around that position. The issue was that as the user scrolled left immediately after that, they'd get past the last image and see a black screen. Further scrolling around would get rid of it. This change, setting the currentIndex in moveToIndex, solves that problem.

We also added a simple label to show the rating assigned to a given meal. This definitely needs to be made prettier.

Lastly, GallleryViewController wasn't initializing its appDelegate variable, making it so that no ratings were actually persisted to the database.