Fixed a timezone bug, show timeTaken for a meal in Debug builds
When we set notifications, we now use the local time zone. So, if the user switches time zones after taking a picture, things still work properly. I ran into this traveling to Austin.

Also, to help with debugging I've added a field to the mealView which only shows up in debug builds - it shows you when you took the picture in question.